pyproject.toml
[project]
name = "eterna-labs-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "fastapi>=0.121.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "httpx>=0.24.0",
    "uvicorn[standard]>=0.38.0",
    "python-dotenv>=1.2.1",
]
